# openapi.yaml
openapi: 3.0.3
info:
  title: QRATEX API
  description: |
    QRATEX müşteri deneyimi ve geri bildirim API'si.
    Çoğu endpoint oturum (session cookie) gerektirir; Bearer token bazı entegrasyonlarda kullanılabilir.
  version: 1.0.0
servers:
  - url: /
    description: Uygulama kökü (relative path)
paths:
...
  /api/user/profile:
    get:
      summary: Kullanıcı profili
      description: Oturum açmış kullanıcının profil bilgisi (ad, e-posta, adres, işletme alanları vb.).
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  user: { type: object }
        '401': { description: Unauthorized }
    patch:
      summary: Profil güncelle
      description: Ad, avatar, işletme adı, adres, dil tercihi vb. güncelleme.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                image: { type: string }
                businessName: { type: string }
                address: { type: string }
                latitude: { type: number }
                longitude: { type: number }
                preferredLanguage: { type: string }
      responses:
        '200': { description: OK }
        '401': { description: Unauthorized }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: NextAuth session cookie veya Bearer token (entegrasyon)